feat(DelimiterExtractor): DelimiterExtractor 가 커스텀 구분자를 제외한 남은 입력도 함께 반환하도록 변경
- ExtractionResult DTO 작성
